% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_auth.R
\name{oktopost_auth}
\alias{oktopost_auth}
\title{oktopost_auth}
\usage{
oktopost_auth(account_id, access_token)
}
\arguments{
\item{account_id}{The Oktopost ID}

\item{access_token}{The Oktopost API token}
}
\description{
Authenticate your Oktopost API account
}
