% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_followers.R
\name{oktopost_followers}
\alias{oktopost_followers}
\title{oktopost_followers}
\usage{
oktopost_followers(profile_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{profile_id}{The social profile ID}

\item{start_date}{The start date you want to check - defaults to -30 days}

\item{end_date}{The end date you want to check - defaults to today}
}
\description{
Get a daily count of social followers by profile
}
