% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_message_list.R
\name{oktopost_message_list}
\alias{oktopost_message_list}
\title{oktopost_message_list}
\usage{
oktopost_message_list(campaign_id, with_tags = NULL, network = NULL)
}
\arguments{
\item{campaign_id}{The campaign ID - mandatory field}

\item{with_tags}{If you want to add your tags or not}

\item{network}{The social network that you want to get messages for. Optional.}
}
\description{
Get a list of all of the messages you've passed
}
