% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_messages.R
\name{oktopost_messages}
\alias{oktopost_messages}
\title{oktopost_messages}
\usage{
oktopost_messages(message_id)
}
\arguments{
\item{message_id}{The post ID}
}
\description{
Get information about an individual message
}
