% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_post_analytics.R
\name{oktopost_post_analytics}
\alias{oktopost_post_analytics}
\title{oktopost_post_analytics}
\usage{
oktopost_post_analytics(post_id)
}
\arguments{
\item{post_id}{The post ID}
}
\description{
Get information about an individual post
}
