% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_post_analytics.R
\name{oktopost_post_analytics_list}
\alias{oktopost_post_analytics_list}
\title{oktopost_post_analytics_list}
\usage{
oktopost_post_analytics_list(post_ids, verbose = F)
}
\arguments{
\item{post_ids}{The post IDs}

\item{verbose}{Add confirmation that each post_id has been processed}
}
\description{
Get information about many posts
}
