% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_posts_list.R
\name{oktopost_posts_list}
\alias{oktopost_posts_list}
\title{oktopost_posts_list}
\usage{
oktopost_posts_list(
  message_id = NULL,
  campaign_id = NULL,
  status = NULL,
  created_by = NULL,
  source = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{message_id}{The message ID}

\item{campaign_id}{The campaign ID}

\item{status}{The post status}

\item{created_by}{Who created the post}

\item{source}{The social source}

\item{before}{Start date}

\item{after}{End date}
}
\description{
Get a list of all of your posts
}
