% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_post_analytics.R
\name{oktopost_social_post}
\alias{oktopost_social_post}
\title{oktopost_social_post}
\usage{
oktopost_social_post(post_id)
}
\arguments{
\item{post_id}{The post ID}
}
\description{
Get information about a social post
}
