% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_post_analytics.R
\name{oktopost_social_post_analytics}
\alias{oktopost_social_post_analytics}
\title{oktopost_social_post_analytics}
\usage{
oktopost_social_post_analytics(social_post_id)
}
\arguments{
\item{social_post_id}{The social post ID}
}
\description{
Get more detailed information about a social post
}
