% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_post_analytics.R
\name{oktopost_social_posts_all}
\alias{oktopost_social_posts_all}
\title{oktopost_social_posts_all}
\usage{
oktopost_social_posts_all(post_df)
}
\arguments{
\item{post_df}{The output of the oktopost_posts_list function}
}
\description{
Get all of the social posts in a single DF
}
