% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oktopost_social_profiles.R
\name{oktopost_social_profiles}
\alias{oktopost_social_profiles}
\title{oktopost_social_profiles}
\usage{
oktopost_social_profiles()
}
\description{
Get information about the connected social profiles
}
