% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{parquet_write}
\alias{parquet_write}
\title{parquet_write}
\usage{
parquet_write(df, file_name = NULL, path = "/tmp")
}
\arguments{
\item{df}{The dataframe}

\item{file_name}{The name of the file}

\item{path}{The filepath}
}
\description{
Dump a dataframe to a parquet file for loading into Snowflake
}
